Move api client model to shared database package
