+++ packages/database/src/modelClasses/index.js
@@ -6,6 +6,7 @@
 import { AccessRequestModel } from './AccessRequest';
 import { AncestorDescendantRelationModel } from './AncestorDescendantRelation';
 import { AnswerModel } from './Answer';
+import { APIClientModel } from './APIClient';
 import { CommentModel } from './Comment';
 import { CountryModel } from './Country';
 import { DashboardGroupModel } from './DashboardGroup';
@@ -45,6 +46,7 @@ export const modelClasses = {
   AccessRequest: AccessRequestModel,
   AncestorDescendantRelation: AncestorDescendantRelationModel,
   Answer: AnswerModel,
+  ApiClient: APIClientModel,
   Comment: CommentModel,
   Country: CountryModel,
   DashboardGroup: DashboardGroupModel,
