+++ packages/meditrak-server/src/database/models/index.js
@@ -5,7 +5,6 @@
 
 export { AccessRequestModel as AccessRequest } from './AccessRequest';
 export { AnswerModel as Answer } from './Answer';
-export { APIClientModel as ApiClient } from './APIClient';
 export { ApiRequestLogModel as ApiRequestLog } from './ApiRequestLog';
 export { CommentModel as Comment } from './Comment';
 export { CountryModel as Country } from './Country';
